Change traversed set to visited set
